+++ kythe/cxx/extractor/cxx_extractor.h
@@ -179,7 +179,8 @@ class IndexWriter {
       std::unique_ptr<IndexWriterSink> sink,
       const std::string &main_source_file, const std::string &entry_context,
       const std::unordered_map<std::string, SourceFile> &source_files,
-      const HeaderSearchInfo *header_search_info, bool had_errors);
+      const HeaderSearchInfo *header_search_info, bool had_errors,
+      const std::string &clang_working_dir);
   /// \brief Set the fields of `file_input` for the given file.
   /// \param clang_path A path to the file as seen by clang.
   /// \param source_file The `SourceFile` to configure `file_input` with.